홍영현 : 인벤토리 더블클릭기능, 사운드, 위치 바꾸기기능, 사운드 추가

+++ WinApi_MapleStory/GameEngineContents/Player.h
@@ -71,7 +71,7 @@ public:
 		PlayerHp += _Value;
 	}
 
-	float GetMainPlayerHpValue()
+	int GetMainPlayerHpValue()
 	{
 		return PlayerHp;
 	}
@@ -81,7 +81,7 @@ public:
 		PlayerMp += _Value;
 	}
 
-	float GetMainPlayerMpValue()
+	int GetMainPlayerMpValue()
 	{
 		return PlayerMp;
 	}
